test passes after function move to own file and exported/imported
